feat: Add windows wsl specific configs - blue colored
Add wsl specific configs for tmux, vim and zsh. Tmux
and zsh are blue colored. Vim is tailored with to work
with Windows terminal, so that all colors and highlights
are well visible. Some vim color schemes are unusable
in Windows terminal - highlight and text the same color
for example.